Rename shared-models to shared/models
