+++ core/models/guild.js
@@ -1,4 +1,4 @@
-const GuildModel = require("./../../shared-models/guild.js");
+const GuildModel = require("./../../shared/models/guild.js");
 const UserRole = require("./user-role.js");
 const Permission = require("./permission.js");
 const Plugin = require("./plugin.js");
