+++ core/models/user-role.js
@@ -1,4 +1,4 @@
-const UserRoleModel = require("./../../shared-models/user-role.js");
+const UserRoleModel = require("./../../shared/models/user-role.js");
 const db = require("./../database/crud.js");
 const Logger = require("../utils/logger.js");
 const discordClientProvider = require("./../utils/discord-client-provider.js");
