+++ dashboard/www/index.html
@@ -66,14 +66,14 @@
     <script src="/dashboard/scripts/utils/mobile-helper.js"></script>
 
     <!-- Models definition -->
-    <script src="/shared-models/bot-settings.js"></script>
-    <script src="/shared-models/command.js"></script>
-    <script src="/shared-models/guild.js"></script>
-    <script src="/shared-models/owner.js"></script>
-    <script src="/shared-models/permission.js"></script>
-    <script src="/shared-models/plugin.js"></script>
-    <script src="/shared-models/repository.js"></script>
-    <script src="/shared-models/user-role.js"></script>
+    <script src="/shared/models/bot-settings.js"></script>
+    <script src="/shared/models/command.js"></script>
+    <script src="/shared/models/guild.js"></script>
+    <script src="/shared/models/owner.js"></script>
+    <script src="/shared/models/permission.js"></script>
+    <script src="/shared/models/plugin.js"></script>
+    <script src="/shared/models/repository.js"></script>
+    <script src="/shared/models/user-role.js"></script>
 
     <!-- Models -->
     <script src="/dashboard/scripts/models/channel.js"></script>
