+++ shared/models/plugin.js
